One last pivot from MAUI to Avalonia
I've tried to make MAUI work but unfortunately, it is just not yet ready for the use cases I want. As such I am moving away from MAUI and instead implementing Avalonia as it is far more stable and will be significantly simpler to implement. Thankfully very little UI was worked on and as such, we can keep using the majority of the work that was already done.