Switching over to .NET MAUI
After a lot of time being away from this project. MAUI has had quite a bit of time to mature and as such I find that it is now in a state that it can be used for this project. Also MAUI handles MVVM and UI far easier than Avalonia which has been a slug fest of trying to understand how to just get UI to work. As such I've decided to abandon Avalonia and move back to MAUI.